+++ client/src/components/Prompts/Groups/List.tsx
@@ -1,12 +1,11 @@
-import { FileText, Plus } from 'lucide-react';
-import { Link } from 'react-router-dom';
-import { Button, Skeleton } from '@librechat/client';
-import { PermissionTypes, Permissions } from 'librechat-data-provider';
+import { FileText } from 'lucide-react';
+import { Skeleton } from '@librechat/client';
+import { Permissions } from 'librechat-data-provider';
 import type { TPromptGroup, TStartupConfig } from 'librechat-data-provider';
 import DashGroupItem from '~/components/Prompts/Groups/DashGroupItem';
 import ChatGroupItem from '~/components/Prompts/Groups/ChatGroupItem';
 import { useGetStartupConfig } from '~/data-provider';
-import { useLocalize, useHasAccess } from '~/hooks';
+import { useLocalize } from '~/hooks';
 import { cn } from '~/utils';
 
 export default function List({
@@ -21,28 +20,9 @@ export default function List({
   const localize = useLocalize();
   const { data: startupConfig = {} as Partial<TStartupConfig> } = useGetStartupConfig();
   const { instanceProjectId } = startupConfig;
-  const hasCreateAccess = useHasAccess({
-    permissionType: PermissionTypes.PROMPTS,
-    permission: Permissions.CREATE,
-  });
 
   return (
     <div className="flex h-full flex-col">
-      {hasCreateAccess && (
-        <div className="flex w-full justify-end">
-          <Button
-            asChild
-            variant="outline"
-            className={cn('w-full bg-transparent', !isChatRoute && 'mx-2')}
-            aria-label={localize('com_ui_create_prompt')}
-          >
-            <Link to="/d/prompts/new">
-              <Plus className="size-4" aria-hidden="true" />
-              {localize('com_ui_create_prompt')}
-            </Link>
-          </Button>
-        </div>
-      )}
       <div className="flex-grow overflow-y-auto" aria-label={localize('com_ui_prompt_groups')}>
         <div className="overflow-y-auto overflow-x-hidden">
           {isLoading && isChatRoute && (
