+++ client/src/components/Prompts/SidePanel/GroupSidePanel.tsx
@@ -4,7 +4,7 @@ import { Button, Sidebar, TooltipAnchor } from '@librechat/client';
 import ManagePrompts from '~/components/Prompts/ManagePrompts';
 import { usePromptGroupsContext } from '~/Providers';
 import List from '~/components/Prompts/Groups/List';
-import PanelNavigation from './PanelNavigation';
+import PanelNavigation from '../Groups/PanelNavigation';
 import { useLocalize } from '~/hooks';
 import { cn } from '~/utils';
