+++ client/src/components/Prompts/components/index.ts
@@ -1,0 +1,1 @@
+export { default as CreatePromptButton } from './CreatePromptButton';
